requirements.txt
discord.py>=2.5.2
asyncpg
requests
python-dotenv
